[Add/#70] 1. TKConversationView 로 이름 변경 2. Asset에 accentColor 추가
