+++ talklat/talklat/Resources/Assets.xcassets/AccentColor.colorset/Contents.json
@@ -1,6 +1,15 @@
 {
   "colors" : [
     {
+      "color" : {
+        "color-space" : "srgb",
+        "components" : {
+          "alpha" : "1.000",
+          "blue" : "0x38",
+          "green" : "0x68",
+          "red" : "0xFF"
+        }
+      },
       "idiom" : "universal"
     }
   ],
