+++ talklat/talklat/Sources/Views/Conversation/TKTransitionTestView.swift
@@ -1,0 +1,18 @@
+//
+//  TKTransitionTestView.swift
+//  talklat
+//
+//  Created by Celan on 10/31/23.
+//
+
+import SwiftUI
+
+struct TKTransitionTestView: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+#Preview {
+    TKTransitionTestView()
+}
